+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeediOSTests/Feed UI/FeedSnapshotTests.swift
@@ -10,14 +10,6 @@ import EssentialFeed
 import EssentialFeediOS
 
 final class FeedSnapshotTests: XCTestCase {
-
-    func test_emptyFeed() {
-        let sut = makeSUT()
-        
-        sut.display(emptyFeed())
-        
-        assert(sut.snapShot(), named: "Empty_Feed")
-    }
     
     func test_feedWithContent() {
         let sut = makeSUT()
@@ -42,10 +34,6 @@ final class FeedSnapshotTests: XCTestCase {
         let storyboard = UIStoryboard(name: "Feed", bundle: bundle)
         let feedView = storyboard.instantiateViewController(withIdentifier: "ListViewController") as! ListViewController
         return feedView
-    }
-    
-    private func emptyFeed() -> [CellController] {
-        return []
     }
     
     private func feedWithContent() -> [ImageStub] {
@@ -77,50 +65,6 @@ final class FeedSnapshotTests: XCTestCase {
                 showRetry: true)
         ]
     }
-    
-    private func record(_ snapshot: UIImage, named name: String, file: StaticString = #file, line: UInt = #line) {
-        let snapshotURL = makeSnapshotURL(named: name, file: file, line: line)
-        let snapshotData = makeSnapshotData(snapshot, file: file, line: line)
-            
-        do {
-            try FileManager.default.createDirectory(at: snapshotURL.deletingLastPathComponent(), withIntermediateDirectories: true)
-            try snapshotData?.write(to: snapshotURL)
-        }
-        catch {
-            XCTFail("Failed to record snapshot with error: \(error)", file: file, line: line)
-        }
-    }
-    
-    private func assert(_ snapshot: UIImage, named name: String, file: StaticString = #file, line: UInt = #line) {
-        let snapshotURL = makeSnapshotURL(named: name, file: file, line: line)
-        let snapshotData = makeSnapshotData(snapshot, file: file, line: line)
-        
-        guard let recordedSnapshotData = try? Data(contentsOf: snapshotURL) else {
-            return XCTFail("Failed to load the recorded snapshot at \(snapshotURL). Use 'record' method before asserting", file: file, line: line)
-        }
-        
-        if snapshotData != recordedSnapshotData {
-            let temporaryURL = FileManager.default.temporaryDirectory.appending(path: snapshotURL.lastPathComponent)
-            try! snapshotData?.write(to: temporaryURL)
-            
-            XCTAssertEqual(snapshotData, recordedSnapshotData, "New Snapshot does not match recorded snapshot. New Snapshot URL: \(snapshotURL), Recorded Snapshot URL: \(temporaryURL)", file: file, line: line)
-        }
-    }
-    
-    private func makeSnapshotData(_ snapshot: UIImage, file: StaticString = #file, line: UInt = #line) -> Data? {
-        guard let data = snapshot.pngData() else {
-            XCTFail("Falied to generate PNG data representation from snapshot", file: file, line: line)
-            return nil
-        }
-        return data
-    }
-    
-    private func makeSnapshotURL(named name: String, file: StaticString = #file, line: UInt = #line) -> URL {
-        return URL(fileURLWithPath: String(describing: file))
-            .deletingLastPathComponent()
-            .appending(path: "snapshot")
-            .appending(path: "\(name).png")
-    }
 }
 
 private class ImageStub: FeedImageCellControllerDelegate {
@@ -147,14 +91,5 @@ private extension ListViewController {
             $0.controller = cellController
             return cellController
         })
-    }
-}
-
-private extension UIViewController {
-    func snapShot() -> UIImage {
-        let renderer = UIGraphicsImageRenderer(bounds: view.bounds)
-        return renderer.image { context in
-            view.layer.render(in: context.cgContext)
-        }
     }
 }
