+++ 3.EssentialFeedProject/EssentialFeedProject/EssentialFeed/EssentialFeediOSTests/Feed UI/snapshot/Feed_With_No_Images.png
